Fix compiler assert when byteswap IL is not needed in DAA
The write and read DAA APIs can take as input a `numBytes` argument
which for read operations determines how many bytes to read and for
write operations determines how many bytes to write out to the source or
target byte array respectively. If the user specifies a value of 1 for
this argument no byteswap operation should be performed. For read
operations (unmarshalling) a value of 1 for `numBytes` means that only
a single byte is loaded from the array, so no byteswap is needed. For
write operations a value of 1 for `numBytes` means value to store is
truncated to 1 bytes in length before being written, so no byteswap
is needed.

In both cases we must ensure badIL is not generated.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>