Add replicas of default build configs with OpenJDK JSR292 enabled
The default build configs are the ones with no suffix and "_xl" suffix.

The default build configs are replicated to create new build configs
with OpenJDK JSR292 enabled.

OpenJDK JSR292 is enabled using the following configure option:
"--enable-openjdk-methodhandles".

The new build configs will be used to test the support for OpenJDK
JSR292 until it is confirmed to be stable. After confirmation of
stability, the support for OpenJDK JSR292 will be enabled in the default
configs, and the new build configs will be deleted.

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>