Cache well-known class chain offsets in JITServer client session
Every AOT compilation using SVM at the JITServer sends a
SharedCache_storeSharedData message to the client in
SymbolValidationManager::populateWellKnownClasses(). Most of these calls
return a pointer to the same existing version of well-known class chain
offsets in the client's SCC (after all well-known classes are loaded,
which normally happens early on during start-up). To avoid these
repeated messages, we cache the most recent client-side pointer to
well-known class chain offsets in ClientSessionData, along with a "key"
identifying this instance of well-known classes: the bitset describing
included classes and the list of ROMClass offsets in the client's SCC.
If the cached key doesn't match, populateWellKnownClasses() send a
message to the client and caches the new well-known class chain offsets
pointer in the client session. This is consistent with the behaviour of
the underlying SCC store operation: a new structure is created only if
the bit mask (used as the SCC key) is different, or if the contents (the
list of ROMClass offsets) are different.

Closes: #11137
Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>