Fix issue in DDR field layout calculation
Currently in ObjectFieldInfo DDR uses the GC object allocation alignment
to calculate if a super class is unaligned or not. The VM does not use
the GC object alignment size to determine this, it uses the object size
increment. Due to this mismatch, DDR was interpreting layout differently
on 4bit shift compressedRefs modes. This issue is fixed by defining
`OBJECT_SIZE_INCREMENT_IN_BYTES` in DDR.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>