Ensure we don't crash when image not present
Refactor the VMSnapshotImpl destructor to skip most of
the implementation if the `_vm` is null. This likely
indicates that we're shutting down from a partially
initialized state - likely due to a missing image
file.

Some additional refactoring to make the code slightly
clearer about which state we're in (restore / snapshot)
and what the expected behaviour is.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>