Deprecate nvvmOpCodeNames in favor of a switch
The `nvvmOpCodeNames` in CodeGenGPU.cpp is not generated via the new
Opcodes.enum `OPCODE_MACRO`. There does not exist a nice mapping
between the OMR opcodes and the LLVM equivalents. As such this table
will always need to be manually maintained whenever we modify the opcode
tables either in OMR or OpenJ9.

This table can however be easily translated into a switch statement
since it is so sparse. We carry out this translation within this commit
to eliminate the need to maintain this table in the future.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>