JDK16 VarHandle Support
1. abstract MethodType accessModeTypeUncached(...)
accessModeTypeUncached accepts an AccessType instance instead of an
AccessMode instance as an input argument.

2. final MethodType accessModeTypeUncached(int index)
directs a call to
abstract MethodType accessModeTypeUncached(...)

3. public abstract VarHandle withInvokeExactBehavior()
and its implementation in concrete classes that extend the abstract
VarHandle.

4. public abstract VarHandle withInvokeBehavior()
and its implementation in concrete classes that extend the abstract
VarHandle.

5. public boolean hasInvokeExactBehavior()
returns true if the VarHandle has invokeExact behaviour. Otherwise, it
returns false.

Fixes: https://github.com/eclipse/openj9/issues/11149

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>