Simplify ArrayStoreCHK evaluator
Previously, the evaluator for ArrayStoreCHK included duplicate code for
performing write barriers. While this did provide one small optimization
benefit (it avoided performing two null checks for a certain kind of
write barrier), it required that a lot of very complex code for
performing write barriers be duplicated and would make changes to this
code needlessly difficult due to the way that support is currently
written.

While it should be possible to reintroduce this optimization in the
future, in the interests of simplicity while the write barrier code is
undergoing some modifications, this code has been deduplicated by having
the ArrayStoreCHK evaluator defer to the awrtbari evaluator to perform
the write barrier.

Signed-off-by: Ben Thomas <ben@benthomas.ca>