Split SH_CompositeCacheImpl::startup into 2 parts
Split SH_CompositeCacheImpl::startup into SH_CompositeCacheImpl::startup() and
SH_CompositeCacheImpl::earlystartup(). The purpose of this is that we will later
move earlystartup() to a much earlier stage for shared cache header access. Early
access of compressed reference and gc policy information in the shared class cache
is needed for the Portable Shared Class Cache feature. For more information please
refer to #11232.

Signed-off-by: Harry Yu <harryyu1994@gmail.com>