Serialize Runtime Assumptions into buffer
1. Loop through the runtime assumption kinds
2. Go through all the assumptions for that kind
3. Call the virtual method serialize on the assumption

The owning metadata is the J9JITexceptionTable which holds a pointer
to the sentinel runtime assumption that is part of the circular
linked list of assumptions that belong to a particular JIT'd body.
We need to store the owning metadata so that when we add the
assumption to the RAT in the restore run, we can also enqueue it
into the circular linked list of assumptions for the same JIT'd body.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>