Rename and consolidate bytesInHeaderAndPayload
Rename bytesInHeaderAndPayload to just bytesInHeader. This name has been
a point of confusion for more than one developer because of the fact
that this API does not return total size of the binary relocation record
(ie the data necessary to materialize the value to relocate/validate +
the offsets at which to apply relocation); it in fact only returns the
size of the total data needed to materialize the value. The reason why
it was named "header and payload" is because the "header" was considered
the fields of the base type TR_RelocationRecordBinaryTemplate, whereas
the "payload" was considered the fields of any extending type.

A (possible) point in favour of this is that the fields of
TR_RelocationRecordBinaryTemplate are used to determine the type of the
binary template, and therefore, how to interpret the fields that follow.
However, because immediately following the binary template header fields
are offsets, it complicates the notion that a complete binary relocation
record contains a header, a payload, and offsets - it begs the question
"what makes the offsets different that it cannot be grouped in with the
payload?". And while there is an answer for this, it adds undue
compexity to an already complex infrastructure.

Therefore, to lower the barrier to entry (as I know this was for me),
this commit goes ahead and renames it.

It also deletes all the different implementations of it, and instead
uses the canonical table of sizes to return the answer. In doing so, it
is now consistent with the codegen.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>