Initialization of common fields of the raw relocation record header
The common fields of all the raw relocation record headers (ie the
fields of TR_RelocationRecordBinaryTemplate) should be initialized the
same regardless of platform. This commit enforces this by refactoring
the appropriate code into a common method. The
initializeAOTRelocationHeader method that was implemented on all
platforms is removed, and replaced with
initializePlatformSpecificAOTRelocationHeader. Instead,
only one version of initializeAOTRelocationHeader now exists, which
performs the common intialization, and then delegates to
initializePlatformSpecificAOTRelocationHeader, which in turn may
delegate to initializeCommonAOTRelocationHeader.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>