Keep splitPostGRA stores in right block for acmp fast path
When splitPostGRA splits a block, it might insert register stores and
stores to temporaries at the end of the original block at the split
point.  When splitPostGRA is called for fastpathAcmpHelper, the original
block contains the acmp helper call that will still need to be moved
into a separate "call" block, while any register stores or temporary
stores that do not involve the call to acmp helper will need to remain
in the original block.  Move any treetops, register stores or stores
for the call to the acmp helper to the end of the original block so that
the trees associated with the call can be split into the new "call"
block without pulling any unrelated stores and register stores into the
new block.

Also, register dependencies that splitPostGRA adds to the end of the
"call" block should always be copied to the ifacmpeq node that's
added to the end of the original block as that represents an edge to the
same "target" block.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>