Add more info to assert message in SymRefTab
In order to debug #9416, it would be helpful to know what values are
returned by `TR_ResolvedJ9Method::definingClassFromCPFieldRef` and
`TR_ResolvedRelocatableJ9Method::definingClassFromCPFieldRef`. As such,
this commit adds the result of those to calls to the message of the
assert being triggered. It doesn't make sense to call
`TR_ResolvedRelocatableJ9Method::definingClassFromCPFieldRef` for
regular compilations so the extra information is only added for AOT
compilations.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>