Refactor how ROM strings are accessed on JITServer
- Removed duplicate code for `getROMString` and `getRemoteROMString`
from `TR_ResolvedJ9JITServerMethod`, and added a more generalizable
implementation in `ClientSessionData`
- Added methods to return ROM strings as UTF-8. If a string is
reacheable from the server and the result will be converted to UTF-8,
`getROMStringUTF8` will avoid conversion to `char *` and back to UTF-8.
- Added `TR_ResolvedJ9(JITServer)Method::fieldOrStaticSignature`to
take advantage of the above change and return signature as UTF-8.
- Added `str2utf8` as a new `JITServerHelpers` method to get rid
of multiple implementations of this method in different files.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>