Remove invalid shrc asserts during shutdown
The asserts are accessing _readWriteProtectCntr without holding the
appropriate mutex, the value can be modified by other threads. The value
can be greater than zero (or 1 in the z/OS case) if another thread has
unprotected the read write area. Checking for equality with a particular
value is invalid.

Removing the assert for _headerProtectCntr at the same time since it's
not providing much value.

Testing shows _readWriteProtectCntr had a value of 1 before
`unprotectHeaderReadWriteArea(currentThread, false)` was called, and
reverts to zero after a delay.

The asserts occur during VM shutdown, which for the shared classes
component happens after the trace engine is shut down. The normal
assertion behavior does not occur. If an assert fails the VM exits with
return code 255 (-1) instead of zero.

Issue https://github.com/eclipse/openj9/issues/6361

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>