Update how Eden size is reported at gc-start and gc-end
Fix how Eden size is reported at gc-start and gc-end
in balanced GC policy. Make it consistent with reporting
MM_HeapRegionManagerVLHGC::getHeapMemorySnapshot().

Reorder how/when taxation threshold and regions' age are updated.
Instead of updating region's age at the very end of a GC increment
we now update it at the end of GC increment but before we report
gc-end. With the update of exportStats on how Eden is reported,
we now have total Eden size as being reported as
getCurrentEdenSizeInBytes() always while free eden size in bytes
contain the correct value, where at gc-start _edenFreeHeapSize is
close to zero and at gc-end it's equal to Eden heap size.

Signed-off-by: Igor Braga <higorb1@gmail.com>