Update read barriers to use the new LoadStoreHandler API
Previously, the code for performing read barriers was using the old
MemoryReference-based API for performing its loads. This API is being
deprecated in OMR, so all code dealing with read barriers has been
updated to use the new LoadStoreHandler API.

Unfortunately, this code cannot be updated in its current form to use
LoadStoreHandler as intended since it unavoidably performs the load
during ICF. Thus, it currently uses generateComputeAddressSequence to
compute the address but does not use generateLoadAddressSequence to
perform the load. This is safe since OpenJ9 is the most derived project
that could possibly overload LoadStoreHandler and emitting unnecessary
sync instructions (while bad for performance) is not incorrect.

This code should be revisited later to make it more idiomatic, but that
is beyond the scope of this particular change.

Signed-off-by: Ben Thomas <ben@benthomas.ca>