Reset the last error code on z/OS in j9shmem.c after omrnls_printf()
omrnls_printf() can change the last error, this occurs when
libj9ifa29.so is APF authorized (extattr +a).
The last error may be set to -108 EDC5129I No such file or directory.
This causes cmdLineTester_shareClassesStorageKey to fail with an
unexpected error code. It also shows the incorrect last error as the
reason the shared cache could not be opened, which is confusing. Reset
the last error to what it was before omrnls_printf().

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>