Reset runtimeAssumptionList field in metadata on restore
In the past, the snapshot would only occur at shutdown. Now that it can
occur during runtime, the runtimeAssumptionList field in the metadata
for a JIT'd body should not be set to NULL. The reason is because if,
after the snapshot point, the JVM continues to run and the body is
reclaimed (because of recompilation or any other read), the assumptions
associated with that body will not be deleted.  This is problematic
because the assumptions are registered against specific instruction
addresses. If, after the a JIT'd body is removed and another is placed
in the location it used to reside in, and some event occurs that results
in the assumptions firing, it will patch code in the new body.

Therefore, this commit does not set the runtimeAssumptionList field to
NULL while serialzing the assumptions, and instead sets it to NULL on
restore.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>