Use fewer compilation threads at client when JITServer is low on memory
When the JITServer starts running low on physical memory, it is
detrimental for clients to have all compilation threads active and
sending requests because it makes JITServer use too many compilation
threads which consume a lot of memory.

At the end of each compilation, server tells the client if it's low
on physical memory. As long as that's the case, client will not
activate a new compilation thread and will keep suspending the existing
ones until only one thread is active.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>