Only patch CP entries with valid mapping
During the constantpool creation process, VM will delete constantpool
entires that are not referenced by any bytecode/cp as an optimization.

This will break the patching logic if any of the patched object is not
referenced. So we only patched entries that have a valid mapping in
the RAM class constantpool.

Co-authored-by: Babneet Singh <sbabneet@ca.ibm.com>
Signed-off-by: Jack Lu <Jack.S.Lu@ibm.com>