Fully serialize ROMClasses sent to JITServer
All the UTF8 strings that a ROMClass refers to (including interned ones)
are stored at the end in deterministic order. Strings that are only
referenced by SRPs stored outside the ROMClass bounds (such as method
debug info stored out-of-line) are excluded. Intermediate class data is
removed (if stored inside the ROMClass) and its SRP is set to 0. SRPs to
out-of-line method debug info are set to 0.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>