Enable FSD for AOT Compilations
This commit enables AOT compilations in FSD mode. However, if fields are
being watched, then AOT compilation is disabled. This is because there
is no benefit to having an AOT body with field watch as it increases
the validation complexity, and in case the fields being watched changes
from run to run, the AOT body cannot be loaded.

This commit also adds FSD to the feature flags; this ensures that if the
SCC contains code that was generated when running in FSD mode, it is
only loaded when the JVM is running in FSD mode (and vice versa).

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>