Disable prefetching for aloadi nodes on 64-bit
Previously, the code in the insertPrefetchIfNecessary method was
expecting that it would only see aloadi nodes on 32-bit. Because of
this, the code would not operate correctly and would cause segfaults
during compilation. This was previously being handled by simply not
calling insertPrefetchIfNecessary under these conditions, however
keeping this behaviour requires that OMR know about implementation
details of OpenJ9's prefetching.

While this code could theoretically be made to work under these
conditions, the code as written simply cannot handle this case. Instead,
a check has been added to prevent insertPrefetchIfNecessary from
operating on aloadi nodes on 64-bit to replicate the old behaviour.

Signed-off-by: Ben Thomas <ben@benthomas.ca>