Handle monitor enter/exit on value based instances in x86
If the monitor object type is value based class, VM
will either issue a warning or throw an exception
based on -XX:ValueBasedClassCheck option.

If the monitor object type is unkonwn, insert a runtime
memory check in monitor enter/exit on the class flag
J9ClassIsValueBased.

If the monitor object type is not value based class,
proceed as how it is handled today.

Value based class and value type are mutually exclusive.

Related to JEP390

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>