JITServer shared ROMClass cache implementation
Stores distinct serialized ROMClasses in an unordered map using a hash
of the ROMClass contents as the key and the ROMClass with a reference
count as the value. This design reduces the time spent in the critical
section when inserting a new ROMClass into the map. Access to reference
counts is synchronized using atomic operations in order to reduce lock
contention by not having to acquire the monitor that protects the map
when releasing a ROMClass that still has references.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>