Refactor method breakpoint hook
The code that compensated all the runtime assumptions that get fired on
a method breakpoint did not have any tracing. Addtionally, it is
conceptually very similar to what has to happen for the assumptions
during class redefinition. Therefore, this commit refactors the method
breakpoint hook to be more consistent with the rest of the runtime
assumption compensation mechanism.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>