Move J9 specific changes for CallSnippet to OpenJ9
Changes to identify the PIC Builder helpers for call from snippet that
was implemented in the S390CallSnippet in OMR should be in OpenJ9. This
commit moves those J9 specific API from S390CallSnippet to
S390J9CallSnippet.
Also in this commit in addition to moving those helpers, contains
changes that uses the Object Format to encode the calls in the snippet.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>