Use Original ROMMethod if FSD is enabled
When a method is breakpointed, the VM will copy over part of the
ROMMethod. Therefore, the naive approach of getting the ROMMethod
from the J9Method could result in a ROMMethod pointer that isn't in the
SCC. Therefore, this commit changes TR_J9VMBase::getROMMethodFromMethod
to call getOriginalROMMethod if FSD is enabled.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>