Fix bug related to sequencing scheme in JITServer
JITServer implements a sequencing scheme to apply CHTable updates
and class unload events in the order they happened at the client.
Description of the bug:
A compilation request can be viewed as having 3 attributes:
1) seqNo; the sequence number for the request
2) criticalSeqNo; Specifies the request that must apply its updates
before this request does so
3) lastProcessedCriticalSeqNo; Specifies the last request that has
already applied its updates

The very first compilation request has the following attributes:
 {seqNo=0,  criticalSeqNo=0,  lastProcessedCriticalSeqNo=0}
This allows the first request to pass immediately because
criticalSeqNo==lastProcessedCriticalSeqNo
The second compilation request has the following attributes:
 {seqNo=1,  criticalSeqNo=0,  lastProcessedCriticalSeqNo=0}
So seqNo=1 depends on seqNo=0, but because
criticalSeqNo==lastProcessedCriticalSeqNo, it can too pass
immediately leading to a potential race condition.

The solution is to start the counting for seqNo from 1, so the
first request will look like:
 {seqNo=1,  criticalSeqNo=0,  lastProcessedCriticalSeqNo=0}
and can proceed immediately
The next request will look like:
 {seqNo=2,  criticalSeqNo=1,  lastProcessedCriticalSeqNo=0}
and must wait for lastProcessedCriticalSeqNo to become 1 before
it can proceed.
The problem is when the client attaches to the server, the server
dies and then another server comes up. This server will have
lastProcessedCriticalSeqNo==0 from the initialization, but the
first request will have some criticalSeqNo=N > 0 and may
wait forever for lastProcessedCriticalSeqNo to reach the value N.
To deal with this aspect, the very first request, which is the one
to create the ClientSession, will set lastProcessedCriticalSeqNo to
its own criticalSeqNo.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>