Revert fix for anchored compressedref node in JProfilingValue
One of the issue, we were facing in JProfilingValue when lowering calls
where a child of compressedref anchornode needs uncommoning as the
original node of commoned up compressedref anchornode could exist before
the call and splitting blocks at the call would make anchornode orphan.
Workaround implemented for this issue was to analyse the nodes before
and after split point to make sure we do not make commoned up
compressedref anchor node orphan and this analysis was done in
JProfilingValue optimization. At the time, fix was implemented,
JProfilingValue was using regular block splitter and it does not have
this analysis. As now, JProfilingValue is using post GRA block splitter
to split the block, this analysis for compressedref anchornode can be
easily moved there. This commit removes the analysis from the
JProfilingValue.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>