Replace deprecated gethostbyname() with getaddrinfo() to avoid spurious crash
gethostbyname() used in openConnection() (which is used in JITServer code) is
known to be non MT safe. We have observed rare crashes in the client JVM,
which are likely due to usage of gethostbyname() as explained in #11415
This commit replaces the usage of gethostbyname() with gettaddrinfo() which is
reentrant and therefore more amenable to multithreaded code.

Fixes: #11415

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>