Support exception event hooks in AOT
At JIT bootstrap, if the compiler detects that an exception event could
be hooked, it prevents AOT compilation. There isn't any apparent reason
for why this should be the case. This commit allows AOT compilation to
occur if an exception event can be hooked. However, if a method was
AOT compiled with the assumption that no exception event could be
hooked then the method isn't loaded.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>