Add extra U_64 data field to Indexable Object Headers
Preliminary work to incorporate extra 64 bit field to
Indexable object headers. In case of contiguous arraylets
the extra field will hold the address of where the data
starts, which is right after the header. In case of
discontiguous arraylets the extra field will hold the
address of:

- The contiguous double mapped region (assuming double map
  is enabled).
- The contiguous sparse-heap region outside the heap (assuming
  sparse heap is enabled).
- If double mapping an sparse heap are both disabled, dataAddr
  will point to the address right adter the header, in this
  case to the first arrayoid.

This is the first phase concerning a series of deliverables.
Here, we introduce the extra U_64 data field along with helper
functions such as getters and setters that will be needed in
the future. Now dataAddr is only introduced and never used.
Once it’s stable we’ll deliver the second phase of changes
that makes use of dataAddr.

Please refer to https://github.com/eclipse/openj9/issues/11438
for more info. This set of changes concern part of the first
action item of such issue.

Signed-off-by: Igor Braga <higorb1@gmail.com>