Remove empty jdk folder after move
When 11220 went through, this caused an issue
on Windows. When a bootjdk is downloaded, we leave
behind the empty jdk folder (eg. jdk8u275-b01).
This causes the git clone to fail. Rather than
adding back the cleanws which is slow, we rm the
empty folder to provice a clean workspace again.

Related #10779 #11220
[skip ci]

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>