Fix resolved trampoline reservation on JITServer
Previously, we didn't have JITServer specific code
for handling method trampoline reservations because we assumed
that all methods can be called without trampolines.
This is true for x86, where code repository is enabled by default,
but leads to crashes on Power.

There are 2 types of method trampolines: resolved and unresolved,
corresponding to the types of method calls they enable.
In baseline JIT compilation (no JITServer, no AOT) we reserve
trampolines during codegen and they are patched during method execution.
However, for AOT and JITServer reservation during codegen would not
work, since compiled code is executed by a different VM that doesn't
have trampolines reserved.
AOT uses 2 relocation types to handle this:

1. For unresolved trampolines, we create a `TR_Trampolines` relocation.
This relocation is also added for JITServer compilations so this case
is already handled.

2. For resolved trampolines, AOT only allows their use (and thus resolved
call dispatch) when SVM is enabled because `TR_ResolvedTrampolines`
relocation requires SVM. This is a problem for JITServer, since non-AOT
compilations do not use SVM and we don't want to use unresolved dispatch
for every call.

This commit keeps track of methods for which resolved trampoline is
needed during compilation on the server and sends the list to the client
at the end of the compilation. Once the client receives compiled method
and allocates code cache, it will reserve trampolines.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>