Handle monitor enter/exit on value based instances in Z
If the monitor object type is value based class, VM
will either issue a warning or throw an exception
based on -XX:ValueBasedClassCheck option.

If the monitor object type is unkonwn, insert a runtime
memory check in monitor enter/exit on the class flag
J9_CLASS_DISALLOWS_LOCKING_FLAGS.

If the monitor object type is not value based class,
proceed as how it is handled today.

Related to JEP390

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>