AArch64: Add dummy constant data snippet
ARM64HelperCallSnippets add a GCMap to the instruction next to the `bl` instruction to the helper,
and if an ARM64HelperCallSnippet is emitted at the end of the method,
the address which a GCMap is added is beyond the range of the method.
To avoid that, we add a dummy ConstantDataSnippet.
Data snippets are emitted after normal snippets, so we can make sure that
a snippet emitted at the end of the method is not an ARM64HelperCallSnippet.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>