Fix the value of the romSize field in ROMClass
Before this change, the calculation of the total size of the ROMClass
added the size of the VarHandle MethodType lookup table twice. As a
result, many ROMClasses on jdk11+ include an unused byte range at the
end because their romSize is bigger than the actual size of the data.

The size of the VarHandle MethodType table is already accounted for in
ROMClassWriter::writeVarHandleMethodTypeLookupTable() during the
mark-and-count phase, and is included in the rcWithOutUTF8sSize field of
the ROMClassBuilder::SizeInformation structure. Then it's separately
added again (via the sizeInformation->varHandleMethodTypeLookupTableSize
field) to the total romSize when it's calculated in ROMClassBuilder.cpp.

This commit fixes the issue and removes the now unused code associated
with the sizeInformation->varHandleMethodTypeLookupTableSize field.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>