Adding Injection of Interfaces and IdentityInterface Interface
Added a new JVM mechanism to inject an interfacae into a class.
- interfaces to be injected must be specified in the injectInterfaces
method in ROMClassBuilder.cpp
- Max number of interfaces defined by MAX_INTERFACE_INJECTION

Added a new interface IdentityInterface which specifies all non
ValueType objects. This interface is automatically injected
into any class which is not a ValueType.

Added testing to check that IdentityInterface is added to
reference types and not abstract classes, valuetypes, or
java/lang/Object.

ComparingCursor.cpp writeSRP function has case for SRP_TO_UTF8
to skip equality check with the classFile.

Added new SRP to OPTInfo to carry information of the number of
interfaced that are to be injected.

Signed-off-by: Oussama Saoudi <oussama.saoudi@ibm.com>

srpkeyproducer changes