Refactor prex arg info computation and propagation
This commit addresses problem described in #11584 to get prex arg info
computation and propagation into a better shape. It does the following

- Move prex arg info computation into inliner util `computePrexInfo`.
`computePrexInfo` will not store the result into `_prexArgInfo` of a
call target such that it can be used more broadly.
- There is also a small improvement to arg info computation, where we
walk the trees backwards to find the value of an auto if the auto is
used as a call argument.
- Move `enhance` and `stronger` from `TR_PrexArgInfo` into OMR as OMR
uses them.
- Implement `TR_PrexArgInfo::getCallTree`, which is to be used to
update call site call tree which is needed in prex arg computation.

These changes will allow uniformed way of prex arg info compuation and
propagation in different stages of inliner.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>