Add new relocation records for inlined methods without guards
Currently, there is a coupling between the relocation of the inlining
table in the J9JITExceptionTable struct and inlined method guards; as a
consequence, if a guard was removed in a way that the AOT infrastructure
was not aware of, the associated inlined method would not get relocated.

This commit facilitates removing the dependence on guards by adding
relocation records for inlined methods that may not have a guard. There
already exists such records for Interface and Virtual methods; this
commit adds records for Special, Static, and Abstract methods as well.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>