Tag classes generated for OpenJDK MethodHandles during resolution
Tag the classes, which are generated for OpenJDK MethodHandles so that
the JIT can identify them. The classes are found using the
MemberName->vmtarget method, and they are tagged using the
J9ClassIsGeneratedForOJDKMethodHandle flag during resolution.

Tagging during resolution == tagging before first usage of the method.

An helper, VM_VMHelpers::methodFromOJDKMHGeneratedClass, is also added
so that the JIT can identify methods which belong to the classes
generated for OpenJDK MethodHandles.

CI builds are skipped because OJDK MHs are currently disabled in OpenJ9.

[ci skip]

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>