Include classloader name and module version in stack trace
Javadoc notes that if the classloader is named and it is not one
of the built-in classloaders: Bootstrap, Platform or System (aka
App), then it should be included when printing StackTraceElement.

The RI follows the javadoc, but seems to make an exception for the
System/App classloader, since it prints `app//` in the stack trace.

Currently, we do not specify the classloader name at all. This
change matches the RI behaviour in printing the classloader name
as appropriate and also matches the RI in printing the module
version info for classes loaded by the Application classloader.

[skip ci]

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>