Call correct print function for Unresolved Call Snippet
In the change that refactors and moves J9 specific part of call snippet
to OpenJ9, change in a print function for Unresolved Call Snippet was
missed and it was calling old print function from OMR which was removed.
This was causing an unexpected behaviour when tracing is enabled. This
commit fixes that by calling the correct version of print function for
S390J9CallSnippet while tracing.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>