Use per-client persistent allocators on JITServer
As described in #11162, a regular persistent allocator doesn't
release freed memory back to the system, which causes the memory
occupied by persistent allocator to remain unavailable for other
allocation types, even if all the clients have disconnected.
If the available memory is low, this leads to memory
exhaustion once a large enough number of  clients have cycled through.

This commit enables the server to create a new persistent allocator
and persistent memory instance for each client.
When a client disconnects and its client data is cleaned,
its corresponding persistent allocator and memory will also be destroyed.
This releases all memory reserved by that allocator back into the system.
We modify the persistent allocator to use virtual memory allocation,
instead of simple malloc, because memory allocated with malloc is not
guaranteed to be released back into the system and could also be
retained internally.

The main interface for using per-client allocation is inside
`TR::CompilationInfoPerThreadBase` and contains the following methods:
- `enterPerClientAllocationRegion(clientUID)`: begins per-client
persistent allocation on the current thread for a client with a given UID.
All allocations/frees that would normally be done globally, are now
rerouted to be done by a per-client allocator, e.g. `PERSISTENT_NEW`,
`TR::Compiler->persistentAllocator()`, `jitPersistentAlloc/Free`, etc.
- `exitPerClientAllocationRegion()`: end per-client allocation on the
current thread. All allocations/frees after this call will be done by
the default global allocator.

While the vast majority of persistent allocations can be strictly
divided into per-client/global allocations, there may be some rare
cases when a data structure is allocated globally but can be freed or
reallocated in the middle of a compilation. These cases need to be
handled manually by either using
`TR::Compiler->persistentGlobalAllocator()`
or by exiting the allocation region, performing the
required memory operation and re-entering the region.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>