Remove jdk11 travis build now linter runs via github actions
The travis build compiles xlinux jdk11 via cmake, and also runs the
linter on the JIT code. The cmake build was useful when cmake wasn't the
default, but cmake is the default for jdk11 now. The linter is run via
github actions, added via https://github.com/eclipse/openj9/pull/11648

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>