Deprecate compilationCanBeInterrupted API
This API should have been used in all places that call the analogous API
`compilationShouldBeInterrupted`, but this was not the case. The
ultimate decision we are trying to make is whether a compilation thread
should be interrupted, so the "can" and "should" queries can all be
folded into `compilationShouldBeInterrupted` API. That is, for the
instances in which a compilation thread cannot be intrerrupted we
should just return `false` in `compilationShouldBeInterrupted`.

This change deprecates the API in question and returns `false` if
the compilation thread cannot be interrupted.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>